+++ build.rs
@@ -97,9 +97,6 @@ fn detect_cfgs(target: &mut Target) {
             }
         }),
         ("small_main_stack", &has_small_stack),
-        ("use_prebuilt_docs", &|_| {
-            env_var("FISH_USE_PREBUILT_DOCS").is_some_and(|v| v == "TRUE")
-        }),
         ("using_cmake", &|_| {
             option_env!("FISH_CMAKE_BINARY_DIR").is_some()
         }),